# Massive performance boost for get<double>.