--- a/include/simdjson/document.h
+++ b/include/simdjson/document.h
@@ -71,6 +71,12 @@ public:
   inline size_t after_element() const noexcept;
   really_inline tape_type tape_ref_type() const noexcept;
   really_inline uint64_t tape_value() const noexcept;
+  really_inline bool is_double() const noexcept;
+  really_inline bool is_int64() const noexcept;
+  really_inline bool is_uint64() const noexcept;
+  really_inline bool is_false() const noexcept;
+  really_inline bool is_true() const noexcept;
+  really_inline bool is_null() const noexcept;
   really_inline uint32_t matching_brace_index() const noexcept;
   really_inline uint32_t scope_count() const noexcept;
   template<typename T>