This PR introduces a much faster get<double> function.

Do...

```
mkdir build
cd build
cmake -DSIMDJSON_GOOGLE_BENCHMARKS=ON ..
./benchmark/bench_dom_api 
```

Here are my numbers using master:

```
-----------------------------------------------------------------------------
Benchmark                                   Time             CPU   Iterations
-----------------------------------------------------------------------------
numbers_scan                            28986 ns        28986 ns        24070
numbers_size_scan                       17668 ns        17669 ns        39631
numbers_exceptions_scan                 37095 ns        37095 ns        18891
numbers_exceptions_size_scan            22905 ns        22905 ns        30558
```

Here are my numbers with this PR:

```
-----------------------------------------------------------------------------
Benchmark                                   Time             CPU   Iterations
-----------------------------------------------------------------------------
numbers_scan                            23594 ns        23595 ns        29831
numbers_size_scan                       14255 ns        14255 ns        49109
numbers_exceptions_scan                 28782 ns        28783 ns        24326
numbers_exceptions_size_scan            14246 ns        14246 ns        49116
```

As an aside, observe how the version without exceptions can be, itself, much faster, irrespective of this optimization.


This not 'pretty' code. I just went and wrote straight code with performance in mind without doing any tweaking, any engineering. 


We might consider recommending that people use the exceptionless API for enhanced performance, I will open a separate issue.