@lemire 2020-04-15
> The int64 and uint64 versions should probably treat uint64_t as the unlikely case, since almost everything resolves to int64.

My experience with successive unlikely is not good. If you make a path as unlikely, and then a subpath as unlikely, you get unlikely(unlikely()) and the code generation can get pretty weird. I am not saying it is not a good idea, but I would want to measure it and/or look at the assembly.