@lemire 2020-04-15
@jkeiser Note: this PR illustrates why you may not want to get rid of the "is<T>" as it can be faster in some cases.