@jkeiser 2020-04-15
> My experience with successive unlikely is not good. If you make a path as unlikely, and then a subpath as unlikely, you get unlikely(unlikely()) and the code generation can get pretty weird. I am not saying it is not a good idea, but I would want to measure it and/or look at the assembly.

Oh, I was talking about when we implement `get<int64_t>()` and `get<uint64_t>()`.