@lemire 2020-04-15
@jkeiser 

Oh. I am doing those right now. I will update the PR in a few moments.

I need to go get some cat medication, and I am going to resolve this PR back when I come back.