@lemire 2020-04-15
We really need to do a better job benchmarking the DOM API.