@lemire 2020-04-15
Regarding type()... I have written a new version of numbers_scan  using type() and it is four times slower... 

```
numbers_scan                            23658 ns        23659 ns        29652
numbers_type_scan                       82781 ns        82782 ns         8444
```

I am investigating... This is much slower than I would expect.