@lemire 2020-04-15
So I am going to let the tests run, and then I will merge. I am still very preoccupied by the performance of the type() function. It is tragic.