@lemire 2020-04-15
Something bad happened with the last commit. Let me see.