@lemire 2020-04-15
Odd. It builds fine on my machine.