@lemire 2020-04-15
It was a name clash I did not notice (is_null).