@lemire 2020-04-16
Ok. Merging.